Improve javascript bundles
Improves the javascript bundling by splitting out grpc clients into
their own independent files which can be lazy loaded when visiting
routes which require them.
