+++ frontend/src/components/BrowserCommandGrid/index.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { Space, Spin, Typography } from "antd";
 import type React from "react";
-import { useGrpcClients } from "@/context/GrpcClientsContext";
+import { casByteStreamClient } from "@/grpc/casByteStreamClient";
 import { Command } from "@/lib/grpc-client/build/bazel/remote/execution/v2/remote_execution";
 import type { BrowserPageParams } from "@/types/BrowserPageType";
 import { fetchCasObjectAndParse } from "@/utils/fetchCasObject";
@@ -15,8 +15,6 @@ interface Params {
 }
 
 const BrowserCommandGrid: React.FC<Params> = ({ browserPageParams }) => {
-  const { casByteStreamClient } = useGrpcClients();
-
   const { data, isError, isPending, error } = useQuery({
     queryKey: ["browserCommandGrid", browserPageParams],
     queryFn: () =>
