+++ frontend/src/components/BrowserDirectory/index.tsx
@@ -3,7 +3,7 @@ import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { Link } from "@tanstack/react-router";
 import { Button, Flex, Space, Spin, Typography } from "antd";
 import React, { useEffect } from "react";
-import { useGrpcClients } from "@/context/GrpcClientsContext";
+import { casByteStreamClient } from "@/grpc/casByteStreamClient";
 import {
   type Digest,
   type DigestFunction_Value,
@@ -144,7 +144,6 @@ const RecursiveDirectoryNode: React.FC<RecursiveDirectoryNodeProps> = ({
 }) => {
   const [expanded, setExpanded] = React.useState(isTopLevel);
   const queryClient = useQueryClient();
-  const { casByteStreamClient } = useGrpcClients();
 
   const { data, isError, isPending, error } = useQuery({
     queryKey: [
@@ -189,7 +188,7 @@ const RecursiveDirectoryNode: React.FC<RecursiveDirectoryNodeProps> = ({
         }
       }
     }
-  }, [casByteStreamClient, data, digestFunction, instanceName, queryClient]);
+  }, [data, digestFunction, instanceName, queryClient]);
 
   const calcWillBePrefetched = (
     currentPathHashes: PathHashes | undefined = pathHashes,
