+++ frontend/src/components/SchedulerStatistics/index.tsx
@@ -2,12 +2,10 @@ import { useQuery } from "@tanstack/react-query";
 import { Link } from "@tanstack/react-router";
 import { Skeleton, Space, Statistic } from "antd";
 import type React from "react";
-import { useGrpcClients } from "@/context/GrpcClientsContext";
+import { buildQueueStateClient } from "@/grpc/buildQueueStateClient";
 import PortalAlert from "../PortalAlert";
 
 export const SchedulerStatistics: React.FC = () => {
-  const { buildQueueStateClient } = useGrpcClients();
-
   const { data, isError, error, isPending } = useQuery({
     queryKey: ["listOperations"],
     queryFn: buildQueueStateClient.listOperations.bind(null, {}),
